pyproject.toml
[tool]
[tool.poetry]
name = "randomstuff"
version = "0.0.0"
description = "randomstuff"
authors = [ "Charles Ponti <code@theponti.com>" ]

[tool.poetry.dependencies]
python = "==3.6.9"
coverage = "^5.0.3"
selenium = "^3.141.0"
numpy = "^1.18.1"
pandas = "^1.0.1"
requests = "^2.22.0"
flask = "^1.1.1"
pytest = "^5.3.5"
grpcio = "^1.27.2"
grpcio-tools = "^1.27.2"
todoist-python = "^8.1"
python-dotenv = "^0.11.0"
flask_sqlalchemy = "^2.4"
psycopg2-binary = "^2.8"
